Move async task definitions to new tasks module
